# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BALE BELEQ NEWS API
 * OpenAPI spec version: 0.1.0
 */

export interface Media {
  id: number;
  filename: string;
  url: string;
  type: string;
  size?: number;
  createdAt: string;
}